scripts: added iso install timeout
When installing an ISO using setup-node-appliance.sh, if the kickstart's
postinstall fails, the installation will never end.
For this reason, this patch adds a timeout, leaving the default at -1
(infinity), but scripts that use this utility may modify this value
using the ISO_INSTALL_TIMEOUT environment variable.
In addition to that, the vm now boots with inst.sshd=1 to allow
inspection of the anaconda logs.

Change-Id: I432e41f58f3c7fa7367f1b9527d7494f88ffd577
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>