Fixed scripts to work for current CentOS Stream/RHEL 9
derive-boot-iso.sh script previously used now deprecated isoinfo.

On a CentOS Stream/RHEL 9 system where the package that provided
this command was removed due to being deprecated, this caused to
produce broken builds.

This patch adds code to detect the situation in which the isoinfo
command is not available, and try the xorriso command instead.

If none of the two commands are available the script will now fail.

Signed-off-by: Lev Veyde <lveyde@redhat.com>